Misc gm_1B03 matches

+++ src/melee/vi/vi0102.c
@@ -173,7 +173,7 @@ void vi0102_8031D000_OnFrame(void)
     vi_8031CAAC();
 }
 
-void vi0102_8031D020(s8 arg0, s8 arg1)
+void vi0102_8031D020(int arg0, int arg1)
 {
     un_804D6F60.unk_0 = arg0;
     un_804D6F60.unk_1 = arg1;
